wip: rm unnecessary components
