+++ src/containers/main/MainView.tsx
@@ -1,7 +1,7 @@
 import { useAppConfigStore } from '@app/store/useAppConfigStore';
-import SidebarNavigation from '@app/containers/navigation/SidebarNavigation.tsx';
-import { Dashboard } from '@app/containers/main/Dashboard';
 import { DashboardContainer } from '@app/theme/styles.ts';
+import SidebarNavigation from '@app/containers/navigation/SidebarNavigation.tsx';
+import { Dashboard } from './Dashboard';
 
 export default function MainView() {
     const visualMode = useAppConfigStore((s) => s.visual_mode);
