+++ src/containers/navigation/SidebarNavigation.tsx
@@ -1,9 +1,9 @@
 import { memo } from 'react';
 import { AnimatePresence } from 'motion/react';
 import { useUIStore } from '@app/store/useUIStore.ts';
-import { SidebarGrid, SidebarNavigationWrapper, SidebarWrapper } from './SidebarNavigation.styles.ts';
-import Sidebar from './Sidebars/Sidebar.tsx';
 import SidebarMini from './Sidebars/SidebarMini.tsx';
+import Sidebar from './Sidebars/Sidebar.tsx';
+import { SidebarNavigationWrapper } from './SidebarNavigation.styles.ts';
 
 const SidebarNavigation = memo(function SidebarNavigation() {
     const sidebarOpen = useUIStore((s) => s.sidebarOpen);
@@ -11,15 +11,7 @@ const SidebarNavigation = memo(function SidebarNavigation() {
     return (
         <SidebarNavigationWrapper>
             <SidebarMini />
-            <SidebarGrid>
-                <AnimatePresence>
-                    {sidebarOpen && (
-                        <SidebarWrapper key="sidebar">
-                            <Sidebar />
-                        </SidebarWrapper>
-                    )}
-                </AnimatePresence>
-            </SidebarGrid>
+            <AnimatePresence>{sidebarOpen && <Sidebar />}</AnimatePresence>
         </SidebarNavigationWrapper>
     );
 });
